Include mandatory permissionsRequired property for each endpoint MODCXINV-45
Not introducing any permissions as would either require clients to be
granted permissions to a specific codex implementation or for the codex
multiplexer to be aware of all implementations. Both of which seem to
undermine the intent of the separation, instead relies on client being
granted relevant inventory storage permissions for subsequent requests.