MODCXINV-70: Support instance-storage 9.0 interface
mod-codex-inventory has new major interface versions:

instance-storage 9.0
holdings-storage 6.0
item-storage 10.0

mod-codex-inventory uses only instance-storage, it doesn't use the other two interfaces.

mod-codex-inventory is not affected because the incompatible change is in the DELETE all APIs only that is not used by mod-codex-inventory.

Only the okapiInterfaces need to be bumped in "requires" section of ModuleDescriptor-template.json